Fixed some issues in MetadataMonitor.
It now also monitors the Documents folder in iCloud. Previously was not doing that.

Cleaned up the code that starts the downloads, to make sure there is no interlacing of async tasks causing issues. Basically, now we get the URLs from the query first, synchronously, and pass the URLs around to async functions. Much cleaner.