Introduced FileMonitor to monitor file changes.
NSMetadataQuery seems to work OK for new files, but doesn’t fire on a lot of file updates.

FileMonitor is an NSFilePresenter, and hopefully will pick up a lot more of the remote changes. It also takes care of resolving conflicts.

MetadataMonitor is not simpler. It just monitors for new files in the cloud, and requests they be downloaded if there are some. It doesn’t monitor other changes, or do any conflict resolution.