Introduced CoordinatedFileManager actor so that file coordination is in the background.
Previously, async methods were used, which do have the effect of scheduling a call for later when accessed with await, however, it is still possible that the file coordinator is called on the main thread, and will actually fully block the thread. To fix this, all file coordination is now on an actor, and should never block an important thread.